backend/requirements.txt
# Core API Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
python-multipart>=0.0.6

# Database & ORM
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
supabase>=2.3.0

# AI/ML Pipeline
langchain>=0.1.0
langchain-community>=0.1.0
langgraph>=0.0.15
chromadb>=0.4.22
sentence-transformers>=2.3.1
ollama>=0.1.1

# Data Processing & Graph Analysis
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
networkx>=3.2.1

# Utilities
httpx>=0.26.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.5
pytest-cov>=4.1.0
